Simple change
My 404 page didn't work at all, becausethe source i linked to it was all messed up. I needed to prefix the source with a forward slash in order for it to search in the root rather than the current directory (e.g noelastrom.se/tools/ble, searches for a 404.js in the tools directory). It was a mistake and I've since learned from my incompetence (lol). document.body was null, as the page hadn't been fully loaded yet, this was too a simple fix. The act of adding a defer to the script fixed it. That was all for this commits changes, thank you for listening.