Calc
Calculator for school project, don't mind the random undertale thing, not signiificant as of now, probably won't hold any significance ever either. It is kind of goofy that it is in python too, but it was for demonstrating purposes, in order to teach a friend the basics of programming in python through pygame. Eh.